requirements.txt
sphinx
click>=6.7
pyyaml
nose2
colorama
behave
mako>=1.0
gitpython>=2.1